v0.3.2 update:  eslintignore; npmignore; prettierrc; tsconfig; README; package.json
add:
example code;

other:
transferring onSearchCrow to a separate file;
rename test
