+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-search-crow",
-  "version": "0.3.1",
+  "version": "0.3.2",
   "description": "This Crow will find everything",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
@@ -15,6 +15,11 @@
     "prepare": "husky"
   },
   "devDependencies": {
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/jest": "^27.5.2",
+    "@types/uuid": "^10.0.0",
+    "@types/node": "^16.18.105",
     "@eslint/js": "^9.9.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -24,9 +29,8 @@
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^11.1.6",
-    "@types/jest": "^27.5.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "uuid": "^10.0.0",
+    "react-icons": "^5.3.0",
     "react-scripts": "5.0.1",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
@@ -41,8 +45,15 @@
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "tslib": "^2.6.3",
-    "typescript": "^4.9.5",
     "typescript-eslint": "^8.1.0"
+  },
+  "dependencies": {
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "typescript": "^4.9.5"
+  },
+  "engines": {
+    "node": ">=14.0.0"
   },
   "lint-staged": {
     "src/**/*.{ts,tsx}": [
@@ -55,22 +66,13 @@
   "keywords": [
     "search",
     "react",
-    "search",
     "react-component",
     "typescript"
   ],
   "author": "J0hnnY2late",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/Johnny2late/react-search-crow/issues"
   },
-  "homepage": "https://github.com/Johnny2late/react-search-crow#readme",
-  "dependencies": {
-    "@types/uuid": "^10.0.0",
-    "@types/node": "^16.18.105",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "uuid": "^10.0.0",
-    "react-icons": "^5.3.0"
-  }
+  "homepage": "https://github.com/Johnny2late/react-search-crow#readme"
 }
