+++ tsconfig.json
@@ -26,6 +26,7 @@
   "exclude": [
     "node_modules",
     "dist",
+    "example",
     "**/*.config.js",
     "**/*.config.mjs"
   ]
